fix(reditor): move cursor test dir
