+++ packages/Reditor/src/Cursor/test/util.test.ts
@@ -1,50 +1,4 @@
-import { cursorUpDown } from '../Shared/util';
-import { act } from '@testing-library/react-hooks';
-import { useRecoilValue } from 'recoil';
-import { cursorColS, cursorLnS, useCursorKeymap } from '.';
-import { renderRecoilHook } from '../Shared';
-
-// FIXME: move
-const useMock = () => {
-  const fs = useCursorKeymap();
-  const ln = useRecoilValue(cursorLnS);
-  const col = useRecoilValue(cursorColS);
-
-  return { ...fs, ln, col };
-};
-
-// FIXME: move
-describe('useCursorKeymap', () => {
-  it('right/left', () => {
-    const { result } = renderRecoilHook(useMock);
-    expect(result.current.col).toBe(0);
-
-    act(() => {
-      result.current.right();
-    });
-    expect(result.current.col).toBe(1);
-
-    act(() => {
-      result.current.left();
-    });
-    expect(result.current.col).toBe(0);
-  });
-
-  it('down/up', () => {
-    const { result } = renderRecoilHook(useMock);
-    expect(result.current.ln).toBe(0);
-
-    act(() => {
-      result.current.down();
-    });
-    expect(result.current.ln).toBe(1);
-
-    act(() => {
-      result.current.up();
-    });
-    expect(result.current.ln).toBe(0);
-  });
-});
+import { cursorUpDown } from '../../Shared';
 
 describe('cursorUpDown', () => {
   it('10 or 20', () => {
